- move modding discord button to general tab - add tip for restoring old conhost console - remove debug log that print discord url
