+++ BepInEx.GUI/Views/DisclaimerWindow.axaml
@@ -9,15 +9,8 @@
 		Title="One Time Only Disclaimer">
 		
 	<Grid Margin="10, 10, 10, 10">
-
-		<Grid.RowDefinitions>
-			<RowDefinition Height="0.9*" />
-			<RowDefinition Height="0.1*" />
-		</Grid.RowDefinitions>
-
-		<TextBlock Grid.Row="0" Text="{Binding DisclaimerText}" FontSize="18" TextWrapping="Wrap"/>
 		
-		<Button Grid.Row="1" Content="Modding Discord" Command="{Binding OnClickModdingDiscordLink}" VerticalAlignment="Center" VerticalContentAlignment="Center"/>
+		<TextBlock Grid.Row="0" Text="{Binding DisclaimerText}" FontSize="17" TextWrapping="Wrap"/>
 
 	</Grid>
